+++ src/test/TestConnectionOptions2dx.h
@@ -1,0 +1,19 @@
+//
+//  TestConnectionOptions2dx.h
+//  Adjust SDK
+//
+//  Created by Ugljesa Erceg (@uerceg) on 14th September 2021.
+//  Copyright © 2021 Adjust GmbH. All rights reserved.
+//
+
+#ifndef ADJUST_TESTCONNECTIONOPTIONS2DX_H_
+#define ADJUST_TESTCONNECTIONOPTIONS2DX_H_
+
+#include <iostream>
+
+class TestConnectionOptions2dx {
+public:
+    static void setTestConnectionOptions();
+};
+
+#endif /* ADJUST_TESTCONNECTIONOPTIONS2DX_H_ */