+++ test/app/Classes/AdjustCommandExecutor.h
@@ -23,7 +23,9 @@
 #include "Adjust/AdjustEventSuccess2dx.h"
 #include "Adjust/AdjustSessionSuccess2dx.h"
 #include "Adjust/AdjustSessionFailure2dx.h"
+#include "Adjust/AdjustThirdPartySharing2dx.h"
 #include "Adjust/test/TestLib2dx.h"
+#include "Adjust/test/TestConnectionOptions2dx.h"
 
 class AdjustCommandExecutor {
 private:
@@ -62,6 +64,8 @@ private:
     void trackAdRevenue();
     void disableThirdPartySharing();
     void trackSubscription();
+    void trackThirdPartySharing();
+    void trackMeasurementConsent();
 public:
     static const std::string TAG;
     AdjustCommandExecutor(std::string baseUrl, std::string gdprUrl, std::string subscriptionUrl);